requirements.txt
config>=0.5.1
keyboard>=0.13.5
mss>=7.0.1
ultralytics>=8.0.0
opencv-python
pyserial
mss
